+++ src/Layout/index.ts
@@ -1,2 +1,2 @@
-export * from "./Layout"
+export * from "./HomeLayout"
 export * from "./MapLayout"
